requirements.txt
streamlit
pandas
pdf2image
pytesseract
Pillow
openpyxl
